+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+- [#13](https://github.com/NoelDeMartin/umai/issues/13) Image upload with long titles on mobile.
+
 ## [v0.1.3](https://github.com/NoelDeMartin/umai/releases/tag/v0.1.3) - 2023-03-17
 
 ### Fixed